functional_call/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=1.10.0
requests>=2.31.0
python-dotenv>=1.0.0
langid>=1.1.6
pymodbus==2.5.3

# 本地路由（BART + AdaptiveClassifier，首次启用会下载模型权重）
transformers>=4.40.0
adaptive_classifier>=0.0.12
torch>=2.0.0
sentencepiece>=0.2.0

